Video: Multithread CPU vertex shaders
Changes to the vertex cache to allow the vertex shaders to run in
parallel. This does not attempt any kind of accuracy in the emulaton
(such as emulation the four shader engines and execution ordering) as
doing so would kill any performance benefits, and doesn't actually
fix any known issues.